getAggregatesCsv in AggregatesController
