+++ apps/onboarding-ms/src/main/docs/openapi.json
@@ -25,6 +25,51 @@
     "name" : "support"
   } ],
   "paths" : {
+    "/v1/aggregates/csv/{onboardingId}/{productId}" : {
+      "get" : {
+        "tags" : [ "Aggregates Controller" ],
+        "summary" : "Retrieve aggregates csv for a given onboarding and product",
+        "description" : "Downloads the aggregates csv associated with the specified onboarding ID and product.",
+        "operationId" : "getAggregatesCsv",
+        "parameters" : [ {
+          "name" : "onboardingId",
+          "in" : "path",
+          "required" : true,
+          "schema" : {
+            "type" : "string"
+          }
+        }, {
+          "name" : "productId",
+          "in" : "path",
+          "required" : true,
+          "schema" : {
+            "type" : "string"
+          }
+        } ],
+        "responses" : {
+          "200" : {
+            "description" : "OK",
+            "content" : {
+              "application/octet-stream" : {
+                "schema" : {
+                  "format" : "binary",
+                  "type" : "string"
+                }
+              }
+            }
+          },
+          "401" : {
+            "description" : "Not Authorized"
+          },
+          "403" : {
+            "description" : "Not Allowed"
+          }
+        },
+        "security" : [ {
+          "SecurityScheme" : [ ]
+        } ]
+      }
+    },
     "/v1/aggregates/verification/prod-io" : {
       "post" : {
         "tags" : [ "Aggregates Controller" ],
@@ -1353,51 +1398,6 @@
                   "items" : {
                     "$ref" : "#/components/schemas/TokenResponse"
                   }
-                }
-              }
-            }
-          },
-          "401" : {
-            "description" : "Not Authorized"
-          },
-          "403" : {
-            "description" : "Not Allowed"
-          }
-        },
-        "security" : [ {
-          "SecurityScheme" : [ ]
-        } ]
-      }
-    },
-    "/v1/tokens/{onboardingId}/aggregates-csv/{productId}" : {
-      "get" : {
-        "tags" : [ "Token Controller" ],
-        "summary" : "Retrieve aggregates csv for a given onboarding and product",
-        "description" : "Downloads the aggregates csv associated with the specified onboarding ID and product.",
-        "operationId" : "getAggregatesCsv",
-        "parameters" : [ {
-          "name" : "onboardingId",
-          "in" : "path",
-          "required" : true,
-          "schema" : {
-            "type" : "string"
-          }
-        }, {
-          "name" : "productId",
-          "in" : "path",
-          "required" : true,
-          "schema" : {
-            "type" : "string"
-          }
-        } ],
-        "responses" : {
-          "200" : {
-            "description" : "OK",
-            "content" : {
-              "application/octet-stream" : {
-                "schema" : {
-                  "format" : "binary",
-                  "type" : "string"
                 }
               }
             }
