+++ apps/onboarding-ms/src/main/java/it/pagopa/selfcare/onboarding/controller/AggregatesController.java
@@ -14,6 +14,7 @@ import lombok.AllArgsConstructor;
 import org.eclipse.microprofile.openapi.annotations.Operation;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.jboss.resteasy.reactive.RestForm;
+import org.jboss.resteasy.reactive.RestResponse;
 
 import java.io.File;
 
@@ -63,6 +64,18 @@ public class AggregatesController {
     public Uni<VerifyAggregateResponse> verifyPagoPaAggregatesCsv(@NotNull @RestForm("aggregates") File file){
 
         return aggregatesService.validatePagoPaAggregatesCsv(file);
+    }
+
+    @Operation(
+            summary = "Retrieve aggregates csv for a given onboarding and product",
+            description = "Downloads the aggregates csv associated with the specified onboarding ID and product."
+    )
+    @GET
+    @Produces(MediaType.APPLICATION_OCTET_STREAM)
+    @Path("/csv/{onboardingId}/{productId}")
+    public Uni<RestResponse<File>> getAggregatesCsv(@PathParam(value = "onboardingId") String onboardingId,
+                                                    @PathParam(value = "productId") String productId){
+        return aggregatesService.retrieveAggregatesCsv(onboardingId, productId);
     }
