+++ apps/onboarding-ms/src/main/java/it/pagopa/selfcare/onboarding/controller/TokenController.java
@@ -55,16 +55,4 @@ public class TokenController {
         return tokenService.retrieveContractNotSigned(onboardingId);
     }
 
-
-    @Operation(
-            summary = "Retrieve aggregates csv for a given onboarding and product",
-            description = "Downloads the aggregates csv associated with the specified onboarding ID and product."
-    )
-    @GET
-    @Produces(MediaType.APPLICATION_OCTET_STREAM)
-    @Path("/{onboardingId}/aggregates-csv/{productId}")
-    public Uni<RestResponse<File>> getAggregatesCsv(@PathParam(value = "onboardingId") String onboardingId,
-                                                    @PathParam(value = "productId") String productId){
-        return tokenService.retrieveAggregatesCsv(onboardingId, productId);
-    }
 }
