+++ apps/onboarding-ms/src/main/java/it/pagopa/selfcare/onboarding/service/AggregatesService.java
@@ -4,6 +4,7 @@ import io.smallrye.mutiny.Uni;
 import it.pagopa.selfcare.onboarding.model.VerifyAggregateAppIoResponse;
 import it.pagopa.selfcare.onboarding.model.VerifyAggregateResponse;
 import it.pagopa.selfcare.onboarding.model.VerifyAggregateSendResponse;
+import org.jboss.resteasy.reactive.RestResponse;
 
 import java.io.File;
 
@@ -14,4 +15,7 @@ public interface AggregatesService {
     Uni<VerifyAggregateSendResponse> validateSendAggregatesCsv (File file);
 
     Uni<VerifyAggregateResponse> validatePagoPaAggregatesCsv (File file);
+
+    Uni<RestResponse<File>> retrieveAggregatesCsv(String onboardingId, String productId);
+
 }
